Default version implemented

+++ source/mkdocs_deploy/shared_implementations.py
@@ -3,12 +3,13 @@ import logging
 import os
 from io import BytesIO
 from tempfile import SpooledTemporaryFile
-from typing import IO
+from typing import IO, Union
 
 from .abstract import RedirectMechanism, TargetSession
 from .versions import DEPLOYMENTS_FILENAME, DeploymentSpec, MIKE_VERSIONS_FILENAME
 
 _logger = logging.getLogger(__name__)
+
 
 def generate_meta_data(deployment_spec: DeploymentSpec) -> dict[str, bytes]:
     """
@@ -24,40 +25,52 @@ def generate_meta_data(deployment_spec: DeploymentSpec) -> dict[str, bytes]:
         MIKE_VERSIONS_FILENAME: deployment_spec.mike_versions().json().encode("utf-8"),
     }
 
+
 class HtmlRedirect(RedirectMechanism):
 
-    def create_redirect(self, session: TargetSession, alias: str, version_id: str) -> None:
-        files_created = set()
-        for filename in session.iter_files(version_id):
-            if filename.endswith(".html") or filename.endswith(".htm"):
-                if filename == "404.html" or filename.endswith("/404.htm"):
-                    session.upload_file(
-                        version_id=alias,
-                        filename=filename,
-                        file_obj=session.download_file(version_id=version_id, filename=filename)
-                    )
-                else:
-                    parts = filename.split("/")
-                    depth = len(parts)
-                    url = ("../" * depth + version_id + "/" + "/".join(parts[:-1]))
-                    session.upload_file(
-                        version_id=alias, # Yes that's correct!
-                        filename=filename,
-                        file_obj=BytesIO(_HTML_REDIRECT_PATTERN.format(url=url).encode("utf-8"))
-                    )
-                files_created.add(filename)
-        for filename in session.iter_files(alias):
-            if filename not in files_created and (filename.endswith("html") or filename.endswith("htm")):
-                session.delete_file(alias, filename)
+    def create_redirect(self, session: TargetSession, alias: Union[str, type(...)], version_id: str) -> None:
+        if alias is ...:
+            session.upload_file(
+                version_id=...,
+                filename="index.html",
+                file_obj=BytesIO(_HTML_REDIRECT_PATTERN.format(url=version_id+"/").encode("utf-8"))
+            )
+        else:
+            files_created = set()
+            for filename in session.iter_files(version_id):
+                if filename.endswith(".html") or filename.endswith(".htm"):
+                    if filename == "404.html" or filename.endswith("/404.htm"):
+                        session.upload_file(
+                            version_id=alias,
+                            filename=filename,
+                            file_obj=session.download_file(version_id=version_id, filename=filename)
+                        )
+                    else:
+                        parts = filename.split("/")
+                        depth = len(parts)
+                        url = ("../" * depth + version_id + "/" + "/".join(parts[:-1]))
+                        session.upload_file(
+                            version_id=alias, # Yes that's correct!
+                            filename=filename,
+                            file_obj=BytesIO(_HTML_REDIRECT_PATTERN.format(url=url).encode("utf-8"))
+                        )
+                    files_created.add(filename)
+            for filename in session.iter_files(alias):
+                if filename not in files_created and (filename.endswith("html") or filename.endswith("htm")):
+                    session.delete_file(alias, filename)
 
-    def refresh_redirect(self, session: TargetSession, alias: str, version_id: str) -> None:
+    def refresh_redirect(self, session: TargetSession, alias: Union[str, type(...)], version_id: str) -> None:
         # create_redirect already cleans up so no need to explicitly delete the old one
         self.create_redirect(session, alias, version_id)
 
-    def delete_redirect(self, session: TargetSession, alias: str) -> None:
-        for filename in session.iter_files(alias):
-            if filename.endswith("html") or filename.endswith("htm"):
-                session.delete_file(version_id=alias, filename=filename)
+    def delete_redirect(self, session: TargetSession, alias: Union[str, type(...)]) -> None:
+        if alias is ...:
+            session.delete_file(version_id=..., filename="index.html")
+        else:
+            for filename in session.iter_files(alias):
+                if filename.endswith("html") or filename.endswith("htm"):
+                    session.delete_file(version_id=alias, filename=filename)
+
 
 _HTML_REDIRECT_PATTERN="""<!DOCTYPE html>
 <html>
@@ -77,9 +90,11 @@ _HTML_REDIRECT_PATTERN="""<!DOCTYPE html>
 </html>
 """
 
+
 SHARED_REDIRECT_MECHANISMS = {
     'html': HtmlRedirect()
 }
+
 
 class SeekableFileWrapper(contextlib.closing):
     """
